+++ src/lib.rs
@@ -1,28 +1,23 @@
 #![doc = include_str!("../README.md")]
+#![no_std]
+#![cfg_attr(target_feature = "atomics", feature(stdarch_wasm_atomic_wait))]
 #![allow(non_upper_case_globals)]
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 
+extern crate alloc;
+
 #[rustfmt::skip]
 #[allow(clippy::type_complexity)]
-mod libsqlite3;
-
+mod bindings;
+mod memvfs;
 mod shim;
 
-/// Virtual File System implementations for different browser storage backends.
-mod vfs;
+/// Low-level utilities, traits, and macros for implementing custom SQLite Virtual File Systems (VFS)
+pub mod utils;
 
 /// Raw C-style bindings to the underlying `libsqlite3` library.
-pub use libsqlite3::*;
+pub use bindings::*;
 
 /// In-memory VFS implementation.
-pub use vfs::memory as mem_vfs;
-
-/// Origin Private File System (OPFS) VFS implementation using `SyncAccessHandle`.
-pub use vfs::sahpool as sahpool_vfs;
-
-/// Utility functions and types to help with creating custom VFS implementations.
-pub use vfs::utils;
-
-#[cfg(test)]
-wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_dedicated_worker);
+pub use memvfs::{MemVfsError, MemVfsUtil};